refactor(Ladder): Ladder interface 추출한후, Abstract 클래스 만들기, 이후 이를 상속받는 RandomLadder, UserDefineLadder 만들기
기존의 RandomLadder, UserDefineLadder 는 중복된 부분이 매우 많았다.
따라서 이를 interface로 추출한 후, 공통된 로직은 Abstract 클래스에서 구현하고, 서로 다르게 사다리를 만드는 로직인 makeLines는 추상 팩토리 패턴을 사용하여 자식 class가 상속하여 구현하도록 하였다.
